Initial work on creating service factories for service objects
