+++ cas-server-core/src/main/java/org/jasig/cas/authentication/principal/AbstractServiceFactory.java
@@ -19,6 +19,9 @@
 
 package org.jasig.cas.authentication.principal;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import javax.servlet.http.HttpServletRequest;
 
 /**
@@ -28,6 +31,9 @@ import javax.servlet.http.HttpServletRequest;
  * @since 4.2
  */
 public abstract class AbstractServiceFactory<T extends Service> implements ServiceFactory<T> {
+
+    /** Logger instance. */
+    protected final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Override
     public <T1 extends Service> T1 createService(final String id, final Class<? extends Service> clazz) {
