+++ cas-server-support-openid/src/main/java/org/jasig/cas/support/openid/authentication/principal/OpenIdService.java
@@ -35,9 +35,6 @@ import org.openid4java.message.ParameterList;
 import org.openid4java.server.ServerManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.util.StringUtils;
-
-import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -170,31 +167,6 @@ public final class OpenIdService extends AbstractWebApplicationService {
     @Override
     public boolean isLoggedOutAlready() {
         return true;
-    }
-
-    /**
-     * Creates the service from the request.
-     *
-     * @param request the request
-     * @param openIdPrefixUrl the prefix url for OpenID
-     * @return the OpenID service
-     */
-    public static OpenIdService createServiceFrom(
-            final HttpServletRequest request, final String openIdPrefixUrl) {
-        final String service = request.getParameter(OpenIdConstants.OPENID_RETURNTO);
-        final String openIdIdentity = request.getParameter(OpenIdConstants.OPENID_IDENTITY);
-        final String signature = request.getParameter(OpenIdConstants.OPENID_SIG);
-
-        if (openIdIdentity == null || !StringUtils.hasText(service)) {
-            return null;
-        }
-
-        final String id = cleanupUrl(service);
-        final String artifactId = request.getParameter(OpenIdConstants.OPENID_ASSOCHANDLE);
-        final ParameterList paramList = new ParameterList(request.getParameterMap());
-
-        return new OpenIdService(id, service, artifactId, openIdIdentity,
-                signature, paramList, openIdPrefixUrl);
     }
 
     @Override
