+++ cas-server-support-openid/src/test/java/org/jasig/cas/support/openid/authentication/principal/OpenIdServiceTests.java
@@ -71,7 +71,7 @@ public class OpenIdServiceTests {
 
     @Test
     public void verifyGetResponse() {
-        openIdService = OpenIdService.createServiceFrom(request, null);
+        openIdService = new OpenIdServiceFactory().createService(request);
         final Response response = this.openIdService.getResponse("test");
         try {
             verify(cas, never()).validateServiceTicket("test", openIdService);
@@ -91,7 +91,7 @@ public class OpenIdServiceTests {
     @Test
     public void verifySmartModeGetResponse() {
         request.addParameter("openid.assoc_handle", "test");
-        openIdService = OpenIdService.createServiceFrom(request, null);
+        openIdService = new OpenIdServiceFactory().createService(request);
         Association association = null;
         try {
             association = Association.generate(Association.TYPE_HMAC_SHA1, "test", 60);
@@ -117,7 +117,7 @@ public class OpenIdServiceTests {
     @Test
     public void verifyExpiredAssociationGetResponse() {
         request.addParameter("openid.assoc_handle", "test");
-        openIdService = OpenIdService.createServiceFrom(request, null);
+        openIdService = new OpenIdServiceFactory().createService(request);
         Association association = null;
         try {
             association = Association.generate(Association.TYPE_HMAC_SHA1, "test", 2);
@@ -151,8 +151,8 @@ public class OpenIdServiceTests {
         request2.addParameter("openid.identity", "http://openid.ja-sig.org/battags");
         request2.addParameter("openid.return_to", "http://www.ja-sig.org/?service=fa");
 
-        final OpenIdService o1 = OpenIdService.createServiceFrom(request1, null);
-        final OpenIdService o2 = OpenIdService.createServiceFrom(request2, null);
+        final OpenIdService o1 = new OpenIdServiceFactory().createService(request);
+        final OpenIdService o2 = new OpenIdServiceFactory().createService(request);
 
         assertTrue(o1.equals(o2));
         assertFalse(o1.equals(new Object()));
