--- cas-server-support-openid/src/test/java/org/jasig/cas/support/openid/web/support/OpenIdArgumentExtractorTests.java
@@ -1,44 +1,0 @@
-/*
- * Licensed to Apereo under one or more contributor license
- * agreements. See the NOTICE file distributed with this work
- * for additional information regarding copyright ownership.
- * Apereo licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License.  You may obtain a
- * copy of the License at the following location:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.jasig.cas.support.openid.web.support;
-
-import static org.junit.Assert.*;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.springframework.mock.web.MockHttpServletRequest;
-
-/**
- * @author Scott Battaglia
- * @since 3.1
- */
-public class OpenIdArgumentExtractorTests {
-
-    private OpenIdArgumentExtractor extractor;
-
-    @Before
-    public void setUp() throws Exception {
-        this.extractor = new OpenIdArgumentExtractor();
-    }
-
-    @Test
-    public void verifyNoOpenIdServiceExists() {
-        assertNull(this.extractor.extractService(new MockHttpServletRequest()));
-    }
-}
