+++ cas-server-support-saml/src/main/java/org/jasig/cas/support/saml/authentication/principal/SamlService.java
@@ -24,10 +24,7 @@ import org.jasig.cas.authentication.principal.Response;
 import org.jasig.cas.support.saml.SamlProtocolConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.util.StringUtils;
 
-import javax.servlet.http.HttpServletRequest;
-import java.io.BufferedReader;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -47,16 +44,6 @@ public final class SamlService extends AbstractWebApplicationService {
      * Unique Id for serialization.
      */
     private static final long serialVersionUID = -6867572626767140223L;
-
-    private static final String CONST_START_ARTIFACT_XML_TAG_NO_NAMESPACE = "<AssertionArtifact>";
-
-    private static final String CONST_END_ARTIFACT_XML_TAG_NO_NAMESPACE = "</AssertionArtifact>";
-
-    private static final String CONST_START_ARTIFACT_XML_TAG = "<samlp:AssertionArtifact>";
-
-    private static final String CONST_END_ARTIFACT_XML_TAG = "</samlp:AssertionArtifact>";
-
-    private static final int CONST_REQUEST_ID_LENGTH = 11;
 
     private String requestId;
 
@@ -87,59 +74,6 @@ public final class SamlService extends AbstractWebApplicationService {
         return this.requestId;
     }
 
-    /**
-     * Creates the SAML service from the request.
-     *
-     * @param request the request
-     * @return the SAML service
-     */
-    public static SamlService createServiceFrom(
-            final HttpServletRequest request) {
-        final String service = request.getParameter(SamlProtocolConstants.CONST_PARAM_TARGET);
-        final String artifactId;
-        final String requestBody = getRequestBody(request);
-        final String requestId;
-
-        if (!StringUtils.hasText(service) && !StringUtils.hasText(requestBody)) {
-            LOGGER.debug("Request does not specify a {} or request body is empty",
-                    SamlProtocolConstants.CONST_PARAM_TARGET);
-            return null;
-        }
-
-        final String id = cleanupUrl(service);
-
-        if (StringUtils.hasText(requestBody)) {
-
-            final String tagStart;
-            final String tagEnd;
-            if (requestBody.contains(CONST_START_ARTIFACT_XML_TAG)) {
-                tagStart = CONST_START_ARTIFACT_XML_TAG;
-                tagEnd = CONST_END_ARTIFACT_XML_TAG;
-            } else {
-                tagStart = CONST_START_ARTIFACT_XML_TAG_NO_NAMESPACE;
-                tagEnd = CONST_END_ARTIFACT_XML_TAG_NO_NAMESPACE;
-            }
-            final int startTagLocation = requestBody.indexOf(tagStart);
-            final int artifactStartLocation = startTagLocation + tagStart.length();
-            final int endTagLocation = requestBody.indexOf(tagEnd);
-
-            artifactId = requestBody.substring(artifactStartLocation, endTagLocation).trim();
-
-            // is there a request id?
-            requestId = extractRequestId(requestBody);
-        } else {
-            artifactId = null;
-            requestId = null;
-        }
-
-        LOGGER.debug("Attempted to extract Request from HttpServletRequest. Results:");
-        LOGGER.debug("Request Body: {}", requestBody);
-        LOGGER.debug("Extracted ArtifactId: {}", artifactId);
-        LOGGER.debug("Extracted Request Id: {}", requestId);
-
-        return new SamlService(id, service, artifactId, requestId);
-    }
-
     @Override
     public Response getResponse(final String ticketId) {
         final Map<String, String> parameters = new HashMap<>();
@@ -147,51 +81,4 @@ public final class SamlService extends AbstractWebApplicationService {
         return DefaultResponse.getRedirectResponse(getOriginalUrl(), parameters);
     }
 
-    /**
-     * Extract request id from the body.
-     *
-     * @param requestBody the request body
-     * @return the string
-     */
-    protected static String extractRequestId(final String requestBody) {
-        if (!requestBody.contains("RequestID")) {
-            LOGGER.debug("Request body does not contain a request id");
-            return null;
-        }
-
-        try {
-            final int position = requestBody.indexOf("RequestID=\"") + CONST_REQUEST_ID_LENGTH;
-            final int nextPosition = requestBody.indexOf('"', position);
-
-            return requestBody.substring(position,  nextPosition);
-        } catch (final Exception e) {
-            LOGGER.debug("Exception parsing RequestID from request.", e);
-            return null;
-        }
-    }
-
-    /**
-     * Gets the request body from the request.
-     *
-     * @param request the request
-     * @return the request body
-     */
-    protected static String getRequestBody(final HttpServletRequest request) {
-        final StringBuilder builder = new StringBuilder();
-        try (final BufferedReader reader = request.getReader()) {
-
-            if (reader == null) {
-                LOGGER.debug("Request body could not be read because it's empty.");
-                return null;
-            }
-            String line;
-            while ((line = reader.readLine()) != null) {
-                builder.append(line);
-            }
-            return builder.toString();
-        } catch (final Exception e) {
-            LOGGER.trace("Could not obtain the saml request body from the http request", e);
-            return null;
-        }
-    }
 }
