+++ cas-server-webapp/src/main/webapp/WEB-INF/spring-configuration/argumentExtractorsConfiguration.xml
@@ -26,20 +26,12 @@
        http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd">
     <description>
         Argument Extractors are what are used to translate HTTP requests into requests of the appropriate protocol (i.e.
-        CAS, SAML, SAML2,
-        OpenId, etc.). By default, only CAS is enabled.
+        CAS, SAML, SAML2, OpenId, etc.). By default, only CAS is enabled.
     </description>
 
-    <!--
-    <bean id="samlArgumentExtractor" class="org.jasig.cas.support.saml.web.support.SamlArgumentExtractor" />
-    -->
-
-    <bean
-            id="casArgumentExtractor"
-            class="org.jasig.cas.web.support.CasArgumentExtractor"/>
+    <bean id="casArgumentExtractor" class="org.jasig.cas.web.support.DefaultArgumentExtractor"/>
 
     <util:list id="argumentExtractors">
-        <!-- <ref bean="samlArgumentExtractor" /> -->
         <ref bean="casArgumentExtractor"/>
     </util:list>
 </beans>
